*_set_key*() optimizations: introduced some parallelism by pre-computing shift counts to get right to the second non-zero bit thereby allowing for pairs of bits to be processed in parallel.
Note that when a pair of characters (of the password which used to reside
at the given bit layer and of the newly set password, respectively) differ
by an odd number of bits, an extra non-differing bit will be processed
speculatively.  This cuts the number of conditional branch instructions
by half thus allowing for much better instruction scheduling and an overall
performance improvement (+10% for LM hashes on PowerPC G5, +1.5% on P3).
Unfortunately, this will also result in a slight slowdown on older
processors.  To be remedied by a subsequent algorithm change (pre-compute
the number of differing bits and branch to the appropriate instruction
sequence with no further conditional branches at all).