When loading hashes for cracking, use the internal representation for eliminating any duplicates and already-cracked hashes.
When displaying cracked passwords, let split() unify the encoding of hashes.
(This only works when the only difference is upper vs. lower vs. mixed case
since we're using a hash table and would not do a comparison against hashes
which look very different.)

Force the encoding of LM hashes that get into john.pot to all-lowercase.