1.6.39.3:
When loading hashes for cracking, use the internal representation for
eliminating any duplicates and already-cracked hashes.  This is consistent
with what cracker.c does when logging cracked passwords.

When displaying cracked passwords, let split() unify the encoding of hashes.
(This only works when the only difference is upper vs. lower vs. mixed case
since we're using a hash table and would not do a comparison against hashes
which look very different.)

Force the encoding of LM hashes that get into john.pot to all-lowercase.

Corrected the handling of break statements with nested loops in the external
mode compiler.