bash completion: a few more improvements
Format names are always converted to lower case, even if
the usage output uses upper / mixed case.
A few other options in addition to --format= now work with
a more relaxed upper/lower case handling.

Since --device is the only option name starting with --d, completion
for -d= and --de: is now similar to the one for --dev= or --device=.