relbench: important fix for duplicate format names in file1
Without this fix, for the first file the last benchmark of each format
would be counted (as was the case prior to my changes), whereas for the
second file the fastest one would be counted (which was the intended
logic for both files.

Now, it is always the fastest one that counts.