benchmark-unify: map "MYSQL" to "MySQL" and "NT v2" to "NT MD4"
Otherwise, only the fastest of both benchmarks of newer jumbo revisions
will be compared with the MYSQL_fast implementation, even if for some
reason the MYSQL_fast implementation is slower than the MYSQL one.
Comparing the fastest of two MySQL implementations of an older jumbo
version with the fastest of two MySQL implementations of a new jumbo
version (> john-1.7.9-jumbo-5) is only fair.
And it also gets rid of the "Only in file 1: MYSQL:Raw" output when
running relbench.

The same logic applies to the mapping of NT v2 to MT MD4.