Adjust LDFLAGS for clang targets
For non-debug targets reuse the default LDFLAGS, to make sure -s is added and libraries under /usr/local/lib and /usr/local/ssl/lib are preferred (like with other Linux targets)

This adds -lz to linux-x86-64-clang (only the 3 other clang targets included -lz) and makes sure all the libraries are specified in the same sequence as for all other Linux targets.

For debug targets, don't reuse $(LDFLAGS), because of -s, but use all the -L and -l options from the LDFLAGS in the sequence they are used for the other Linux targets.