Added prepare() format method. Pass a pointer to the format itself to init(), prepare(), valid(), split().
We're not making much use of this functionality in the core JtR tree, but the
equivalent functionality was introduced in -jumbo by JimF and it is being made
use of there.  Hence, this commit lets us reduce the number of changes to core
code in future -jumbo versions, and it lets us proceed with further format
interface revisions in core without having both of the trees deviate from a
common base.