New --list=[sections|parameters=SECTION|list-data=SECTION]
--list=sections prints all existing sections from john.conf
or the config file specified with --config=, resolving includes
The order in which these sections are listed is reversed to
the order of the definition of these sections

--list=parameters:SECTION prints all the parameter definitions
that exist in a section, included sections and included files
are resolved.
If a parameter is defined multiple times, it appears multiple times
in the list.
The order of definitions printed is reversed, compared to the order
of the parameter definitions.

--list=list-data:SECTION prints all the lines in a section that
doesn't contain parameter definitions, but list data.
Included sections and included files are resolved.
Comments (lines starting with '#' or ';') are not printed.
Indention is not preserved, i.e., there are no leading spaces.

--list=help:parameters lists the names of sections which include
parameter definitions

--list=help:list-data lists the names of sections which
(are supposed to) contain list data instead of parameter
definitions

The bash completion script also needed a few adjustments to get
this working.