Dynamic fixes, from Jim.
Here is the patch which should 'fix' both issues. It contains the changes to
dyna.conf  (with x86 lengths at 80-32).  There was bugs in the loader. It
chopped any pw length over 55 down to 55 (for non mmx builds).  It 'should'
have been 80 chars. Actually it 'could' have been 128-9, but I put the length
at 80, and there are quite a few things that were banking on that.