1. in formats.c, the 2nd loop in the 'max length' check, should start at 0, not at 1.
2. there was a crash that happened due to the max length check, in dynamic.
It ONLY happened on dyna_26, and only on 32 bit, and only if not using
SHA_PARA (i.e. is falling back to using the older sha1.S asm code). The bug
was triggered by adding keys in this order:  1, 0, 2, 3.  What happened is
the real key was added first (index 1), then the long keys were added, to
idx 0, 2, 3.  Well, for this specific 'build' of SHA1 code, there was still
a straggler if(index==0) lengths=0   i.e. older code that was before
addition of clear_keys() method.

3. a change to john.c so that -format=dynamic  again 'works' like it did
before the dynamic format was split up from being a single polymorphic
format, to a bunch of 'thin' formats.  Now, -format=dyamic will load all
dynamic formats (very useful in -testing)