formats.h:define FMT_MAIN_VERSION earlier than before
The definition has been moved more to the top of formats.h.
to allow using #if FMT_MAIN_VERSION > 11 ... #endif
for the definition of fmt_main components inside formats.h,
and not just inside format definitions.
You can also use #if FMT_MAIN_VERSION > 11 so thatjohn's
logic, e.g. availability of certain options, can be altered
depending on format version.
New options could be suppressed if the older format definition
does not include the components that are required for the new
optiion to work.
When developing new features, this allows to move most of
the plugin formats to unused, adjust the few remaining formats,
implement the core functionality, revert FMT_MAIN_VERSION to the
older version, restore all the formats previously moved to unused...

It would be good if this trivial change could be applied to core
as well.