OpenCL: Allow auto-tune with a single kernel-invocation limit instead of total crypt_all() limit.
If max_run_time < 1000 it will be parsed as single-invocation limit in ms.
Else, it will be parsed as crypt_all() total max time in ns, like before.

Drop the hard-coded 200 ms limit for AMD, and instead let user specify one in john.conf
if he wish to. Globally or per format.