maint: clean-up source tree
The source tree is "disorganized" with hundreds of "source files in a single
directory". For a starter, this is moving all OpenCL files (and nothing
else) to jumbo/opencl and jumbo/opencl/include.

This patch only moves the OpenCL files and headers. No infra to handle the
new location. The goal is to simplify review.