Revert formats that currently crashes on --tune=auto, see #3091
Revert IPB2 to not using omp_autotune (for now). There's something weird
with it. See #3091.

Revert sapB to not using omp_autotune (for now). There's something weird
with it. See #3091.

Revert SAP-G to not using omp_autotune (for now). There's something weird
with it. See #3091.

Revert XSHA to not using omp_autotune (for now). There's something weird
with it. See #3091.