lotus5-opencl: Always provide the kernel with all inputs
We were sometimes invoking the kernel on more work items than we were copying
inputs for, such as when called with *pcount 1, which got "upgraded" to LWS for
the kernel invocation but not for the inputs copying.  Apparently, when the
kernel ran on some uninitialized inputs it sometimes failed (seemed to depend
on prior content of GPU memory) and the failure was visible (as mysterious
CL_OUT_OF_RESOURCES on NVIDIA) when trying to copy the results back (even the
fewer results for which inputs were provided).

Issue #4343