unique: Hash function update for 64-bit
This corrects support for larger hash tables, and also speeds up a test
case at "-buf=8 -hash-size=29" from 37.2 to 36.6 seconds.

The variable shifts inside the "if (hash & HASH_MAX)" block are
theoretically slow and it is possible to replace them with fixed shifts
perhaps by 32 on 64-bit, but somehow doing so didn't result in a speedup at
least on Sandy Bridge (no slowdown either).  Leaving them as-is for now not
to clutter the code with conditional compilation.

Fixes #4349