crk_remove_hash: Always update the bitmap and hash table
When removing the last hash for a salt, we didn't bother updating the
bitmap and hash table for the salt.  Now we will (unfortunately, at some
increase of memory consumption with "--fork" because of copy-on-write).

We need to do this to avoid logging a guess for the same hash more than
once in case its correct password is tested more than once in the same
crypt_all() call.  Such duplicate logging was not only confusing; it also
resulted in incorrect values of password_count and guess_count.

Fixes #4390