Self-test: require FMT_TRUNC if any over-long test plaintexts exist
Assume that we only include such over-long test plaintexts because the
format has natural truncation at plaintext_length and we want to test that.
Under that assumption, we should then want to tell the rest of JtR that
the truncation is natural rather than an implementation shortcoming, so
that longer candidate passwords are tested (and correctly truncated by the
format's implementation) rather than wrongly skipped (as unsupported).