descrypt, LM: Compare against loaded hashes in crypt_all()
We only do this when there's no bitmap for the current salt since we would
otherwise not make proper use of the bitmap.

This implementation is currently limited to modern x86-alikes with SSE2,
SSE4.1/AVX, AVX2, or MIC/AVX-512, and to scalar builds with OpenMP.
There's no implementation for non-x86 with SIMD yet, nor for x86 with MMX.

Further, we might want to adjust the loader's threshold so that when we do
have this new functionality compiled in, a bitmap would only be created for
higher hash per salt counts than before.  Such adjustment hasn't been made
yet, meaning that this new functionality is made use of in fewer cases than
optimal.  Moreover, the loader's threshold isn't yet aware of our use of
OpenMP here and in cmp_all() vs. our non-use of OpenMP in cracker.c's loop
that uses the bitmap (but we might choose to fix the cracker than than the
loader's threshold in this respect).

See #4442