Introduction of Unicode substitution rules
These are good rules on sites where a user ID is forced to be written as
an ASCII string while the real user name or firstname would have an
alternative Unicode form.  An optimization is used by overwriting the
ASCII character with the second byte of Unicode and then inserting the
first one For each letters, we attempt a replacement of first, second,
and third occurrences and up to all three occurrences.
In order to handle complex cases with multiple substitution, for example
the Icelandic firstname adalradur would have the d's substituted with
eth and the third a with acute, you could use:
-rules-stack=Unicodesubstitution in addition to
-rules=Unicodesubstitution
to combine several substitutions.