Improve handling of UTF-8 rules w/ internal codepage
The first iteration (d595de99) could not handle the case where a
conversion fails because the target codepage lack some needed
character.  This new code does not touch any rule unless it starts
with -U.  After conversion it checks for success and otherwise turn
the -U into -- which is an "always reject" flag (new, and undocumented
in doc/RULES because, well, it's of no use except for this).
Closes #4437.

Also drops a hack for "best-effort" codepage conversion that hardly ever
did any good - it would break the new (and coming) heuristics.