Put a fork in the pkzip format (and zip2john) CS size & source
After scrutinizing specs and open-source code, settle with a hopefully
correct logic for chosing whether to use one or two byte early-rejection,
and whether to use part of the timestamp or of the CRC-32 for it.

This version does not introduce yet another hash input format.  Instead it
uses the original one, filling it correctly.  In order to really maximize
early-rejection we'd need to define a new format that defines 1-or-2-byte
CS at file level as opposed to archive level.  But that may be overkill,
it's not very common to see both.

Closes #4571 (again)