Replace a recently added self-test in pkzip, bc reasons
Due to hysterical reasons too confusing to try to explain, pkzip format
may become subtly confused when given hashes from same archives but
processed with different versions of zip2john.  I gave up trying to sort
it out - it's too odd and too unlikely to be a problem IRL. This is
mostly just a work-around for the Test Suite.

Closes #4578