options.c: Allow --test with --skip-self-tests
While the sound of it is confusing, it really means we allow a
*benchmark* while skipping the self-tests.  This was accidentally
disallowed by the recent review of compatible options combinations.

Also add a sensible error for "--test=0 --skip-self-tests" as that is
a contradiction in terms.  I was sure we had that check at some point
but somehow it wasn't there even looking in older code.

Closes #4601