OpenCL SHA-2: Reintroduce rotate-by-8 workaround for AMD Catalyst
This workaround was dropped in 0c8a15685aad5214c53b4634170838318698f841 and
is now being reintroduced for Catalyst only (not for newer AMD software).

This commit reduces the number of formats failing self-test on a certain
old system with Tahiti and Catalyst from 22 to 13 (plus 1 for rar-opencl,
which segfaults) out of 90 total.

See #4670