OpenCL LWS auto-tune: Bail out if new size was way slower
If speed was worse than half, we don't expect even larger LWS to be
any better.  This was added as a workaround for a POCL problem where
LWS=8 would be fine at 400 ms but all larger sizes taking 11 seconds
or more each, leading to an extremely slow auto-tune for no benfit.

See #4684