OpenCL: Avoid/document strict aliasing violations in shared code
For memcpy, just use void* and char* and start using restrict, hoping
for build-time optimizations and unrolling instead of using rogue code.
Use aliases or memcpy for fixing a bunch of other violations.
A few more or less unrelated optimizations as a bonus, as I found the
opportunities while reviewing.

See #4031