Bug fix: Pot sync would not reset a negative --max-candidates count
Very similar to 96d8ce69 but somehow wasn't caught while at that:
A feature of --max-candidates=N is that if N is negative, the count
resets on any crack.  I realized this must happen for pot syncs as
well, or a fork/MPI job would end up with some nodes exiting much
earlier than others.