Single mode: Bug fix for when to call clear_keys()
I added a call in fb051f3b after seeing it was sometimes missed, but it
led to many double calls and even a fatal situation in the end of a single
run because it would end up with a call to crypt_all() with the key buffer
in limbo.

This fix should do the Right Thing in all situations.