--regen-lost-salts option: Bug fixes, closes #4517
Since every salt has a copy of the initial salt->list pointer, a special
fixup is needed after removing the first binary in the list in
crk_process_guess().  This was the actual #4517 issue.

Other fixes:
The generated salts written to the salt database lacked some info and the
total number of salts wasn't updated.  From now we copy the whole initial
database struct, avoiding several current and at least some possible future
problems.
Improved the "Remaining xx hashes with yy different salts" output to be
more accurate/informative when this feature is used - correctly showing
same-salt boost or salt BF penalty.
Error messages referred to a non-existing file in a non-existing directory.
Several comments were bogus, eg. referring to code that was long gone.
General code-style fixes (including a couple of barely related, made while
trying to understand the code paths involved).
Add bodges for --show to work with hashes cracked using this feature. For
this to work you need to supply the --regen-lost-salts option with --show
too.
Since we unfortunately give the exact format as one of the parameters to
--regen-lost-salts (a fairly stupid decision made 8 years ago), add code to
automatically infer that format whenever a --format parameter isn't given,
or complain if it is and they don't match.
Minor updates to doc/Regen-Lost-Salts, eg. dropping references to things
that were deprecated eight years ago.