Allow hashcat's '+N' rule (increment character) in more situations
We already allowed the '+N' rule without explicitly being in "hashcat mode"
as long as we weren't processing word pairs.

This change allows '+N' at any time where it wouldn't be valid as the
original '+' concatenate word pair command:  If neither of commands "1"
or "2" was used before it, the '+' is parsed as the hashcat command.