7z: For the reported data length "cost", consider early rejection
Now called "data size penalty" instead, to better reflect this.

This is much more usable for actually filtering out hashes: Tens or
hundreds of megabytes without a single byte of padding is a huge hit
to performance (especially for the GPU format) but as soon as we have
a byte or two of padding, the impact very quickly approaches zero.