configure: Bail out if 'perl' is not found
We shouldn't really rely on it, but we currently do for preparing some
dynamic format magic.  All other uses are (to my knowledge) optional,
such as adding Makefile dependencies using plugin_deps.pl, which is nice
but not need.  OTOH nearly all systems have perl, and some of our *2john
tools are written in perl as well.

Closes #4856