Tweaks for knowing NVIDIA GPU type
Drop our notion of Volta (we also lack Turing and Ampere) because it's
unused and currently not needed.  Instead add a notion of "Maxwell or
newer" which is indeed what we normally want.

This speeds up DEScrypt-OpenCL on Pascal and newer GPUs with 65-70%, since
we only checked for "Maxwell" until now and ended up using a suboptimal
kernel for newer GPUs.

See #4897