tezos-opencl: Only transfer keys up to length 48, not 110
48 was the maximum supported in the code anyway, but 110 was also used.

Also, limit mnemonics to a maximum of 128 characters in valid() (was 512)
and use that for the allocation size in the CPU format.  The OpenCL format
had it that way anyway (in the host code, and at effectively 112 in the
device code - now increased), and could probably be made to crash due to
the higher limit in valid().

For the English BIP-0039 word list the worst case is 15*8+14 = 134, and
that doesn't fit.  With a limit of 128, we're betting that the average word
length in a mnemonic isn't close to maximum.  This is OK'ish as long as we
refuse to load the (hopefully non-existent or rare) "hashes" where this
assumption is violated.