nt-opencl: Rebuild kernel on reset() when loaded hash count changes
The kernel is customized to hash table parameters that depend on the
initial loaded hash count.  Without this fix, when not using a mask and
having done self-test, we'd end up running with the self-test kernel
built for the number of hashes in the test db, which wouldn't match the
hash tables built for the actual loaded hashes.  This resulted in false
negatives.

Fixes #4964