Ensure we limit GWS to what the internal mask multiplier allows
This stops us from ending up with a KPC larger than 32-bit.
With the fastest current GPUs we're getting close to this ceiling
and were possibly even breaking it, as the shared code partially failed
to honor it.
The changes also ensure we honor gws_limit for other reasons (such
as memory limitations) in all situations.