mask: change error to warning if ?w missing in hybrid mask
during contest, it's often convenient to change command line from
a previous hybrid mask to a mask not containing ?w, expecting the
wordlist argument to be just ignored. but the error forces one to
remove the wordlist argument (often via ssh, with high lag for
every keypress) and start again, just to add the wordlist later
again if the attempt didn't produce the expected result.
changing it to a warning instead improves usability.