ZIP-opencl: Recover from failed pinned allocations
If we can't have it pinned (a.k.a. page locked), use a normal buffer
instead.  We might want to implement this in more/all formats (many of
Sayantan's already have it).

I opted not to even warn about this - I doubt the format really need pinned
memory anyway and on top of that, the problem I was investigating ended up
something completely different (trying to set up a GWS of zero).