ZIP-opencl: Tweak for autotune with very large data size
This is admittedly a bodge, but it fixes severe problems for archives
with data sizes of hundreds of MB.  During autotune, we now limit the data
size (for final HMAC if early-reject tests passed) to 16 MB.  Without this,
the total crypt_all() duration limit would cause autotune to end up with
far too small figures (sometimes even crashing due to GWS somehow ending up
as zero).

One of the problems is/was that during autotune, we never reject all of a
batch of candidates early.  In real runs, we do.