Perfect hash tables' bitmap bugfixes
The bitmap host code for 4/8-level bitmaps (only 8-level in 128-bit code)
differed from GPU side's: If we mask stuff off we have to do it on GPU as
well or we'll get false negatives.  Never doing it at all yields better
characteristics for "oversized" masks, so that's what we go with.

This bug never surfaced until recent 6ed33a7f, and even then it would
only trigger for NT-opencl, and when having between 10101 and 20100
hashes loaded AND running on either AMD VLIW4 GPU devices, or CPU OpenCL
devices.