configure: Pass default value to AC_CHECK_TOOL, support /usr/ccs/bin/ar
If a tool was not found the variable was left empty, which resulted in
bizarre errors during build.  This change preserves the tool's name in
those failed commands.

Use AC_PATH_TOOL to search for ar in /usr/ccs/bin (as needed on Solaris 10)
and retain this full path in the variable.