Add krb5tgs-sha1 formats: Attacks TGS-REP for etypes 17 and 18
Also known as aes-cts-hmac-sha1-96.  The existing krb5tgs formats are for
etype 23 (rc4-hmac) and too different from 17/18 that it would be relevant
to add this functionality to them.

Also add "cost" notion of etype for a few existing formats, and changed
their benchmark to etype 17 instead of some unknown mix of etypes.

Closes #2809