john.conf: Update the All rule set
While OneRuleToRuleThemAll was a subset of the hashcat rules, it appears
that its revision in OneRuleToRuleThemStill no longer is - it cracks
significantly more hashes in a certain test.  So we should include it along
with hashcat rules in All.

KoreLogic is 7M rules out of a total of 11M (after the changes here), so it
makes sense to have it last.  Some of its rules are very effective on their
own, but most of the passwords they crack are also cracked by non-KoreLogic
rules.

Closes #5307