relbench: Fix calculation of median values
We used simple Perl "sort" on 4 lists for the purpose of calculating and
reporting various median values.  This sorts alphabetically and not
numerically.  Unless all of the numbers in the arrays are of same string length
and with decimal point (if any) at same position, this may give incorrect
results for the medians.

Fix this by changing those 4 lines to use "sort {$a <=> $b}".

Fixes #5367