SIMDSHA512body(): Add redundant initialization of w[] with SSEi_LOOP
This avoids uninitialized warnings with RHEL6's old gcc, and is hopefully
optimized out by more reasonable compilers.  This new code could also be
used without SSEi_LOOP, but trying to do so causes a major performance
regression for e.g. sha512crypt with same RHEL6's old gcc, so we
continue with memcpy() in the else path for now.