logger.c: Retry file lock also after ENOLCK
This means "No locks available".  We now handle it just like
EAGAIN, retrying after a random delay and hoping for recover.
Also adds diagnostic output of the total delay (for both cases)
after this happened, helping in tracking down the problem.

The srand(NODE) is made unconditional and moved to to john.c.
Without it, all nodes are seeded with 1.