Wordlist, rules: Drop the dupe check against the previous word
This check dates back to 1990s when our primary target was descrypt with
its length truncation, its computation was much slower, and we didn't have
the generic dupe suppressor.  In the suppressor, a word matching the
previous one would be an L1 cache hit, so we don't gain much by
pre-checking for that special case.  And if the suppressor is disabled,
that's probably because we don't want any overhead like that.

Besides, this check didn't behave correctly along with an external filter()
possibly modifying the word (a bug that also dates back to 1990s).

OTOH, we do not currently support the suppressor on top of stacked rules
(we do before stacked rules), so we do lose a bit of functionality with
this change.