OpenCL: Mute stupid warnings from nvidia
This suppresses messages like "(): Warning: Function zip_final is a
kernel, so overriding noinline attribute. The function may be inlined
when called." unless verbosity is bumped.  A configuration is added,
"MuteBogusWarnings = Y", which can be toggled for not muting.

We take care to only mute the exact message (bar the function name).
Closes #5456.

Two other configuration options added, but default to old behavior:

NvidiaShowPtxas = Y

AlwaysShowBuildWarnings = N