Huge pages support: Do use them even without MAP_HUGE_2MB
On old systems like RHEL6, do use huge pages if available even at the risk
of failing on munmap() if the kernel is reconfigured to use a non-default
huge page size.  In upstream yescrypt, this tradeoff is resolved the other
way around because failure of a critical system library like libxcrypt is
deemed unacceptable.