OpenCL SHA-512 "free" formats: Fix double race conditions
Only gid 0 cleared the "result" buffer, then a global barrier. The barrier
only synchronizes within a single workgroup so this is not safe. The only
safe and sensible way of clearing it is on host side.

Then multiple threads (potentially) wrote to the same "result" location
without atomic operations. This is a data race and result is undefined in
OpenCL. We have to use atomic operations.