OpenCL inline PTX for nvidia: Tweaks/experiments.
Experiments with mov.b64.
Optionally add "volatile" to inline PTX.
Fix the experimental lut3_64 function so it's actually safe (and
definitely usable, although I see very little changes trying it).

All the above is disabled/unused as-is, so this commit doesn't change
anything: We just get things to experiment further with.

Also change to "n" instead of "i" for immediates. This is what nvidia
specifies for that.