Unify MD4 and MD5 raw formats (CPU and OpenCL)
Unify as in make almost equal - we could also create a md4_md5_common.h
and stuff, but that's too much work.

The MD4 ones were initially very close to copies of MD5 with s/MD5/MD4).
Over the years they always start to differ unneededly and I use to
re-unify them now and then, so a bugfix or improvement in one isn't
missing in the other.

This time we got added support for "{MD4}<base64>" input format.