Document Python modules installation
Amend requirements.txt and move it to run directory, as previously
discussed in #5475.  Add everything needed for being able to run all
of our python tools, including instructions in doc/README-UBUNTU.

All modules exist as Ubuntu packages except protobuf, pyhanko and
pytsk3, so add instructions for that as well as and alternative.

Also fixes permissions on a few python scripts.